+++ src/gameobjects/game/tile.py
@@ -1,7 +1,5 @@
 """
-Represents a single cell in pygame. I could have inherited many
-functions to this class from "Button" but that did not feel sensible.
-After all, this is a game object, not a button   object.
+Represents a single cell in pygame. 
 """
 
 import pygame as pg